requirements.txt
python-dotenv
flask
Flask-SQLAlchemy
flask-login
Flask-Migrate
psycopg2-binary
gunicorn
pytest
requests